refactor: LottoCount 객체 생성
